+++ packages/webpack-bundler-runtime/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @module-federation/webpack-bundler-runtime
 
+## 0.19.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @module-federation/sdk@0.19.1
+  - @module-federation/runtime@0.19.1
+
+## 0.19.0
+
+### Patch Changes
+
+- @module-federation/runtime@0.19.0
+- @module-federation/sdk@0.19.0
+
 ## 0.18.4
 
 ### Patch Changes